moved argparse_to_command cli code into module and configured entry point for it for easier end-user use
