+++ .gitignore
@@ -7,6 +7,7 @@ dist/
 *.egg-info
 .coverage
 coverage.xml
+tmp_*/
 
 # Ignore all directories/files under docs/ unless excluded here
 docs/*/
