+++ entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+argparse_to_command.py = cli_command_parser.conversion.cli:main
