+++ setup.cfg
@@ -5,6 +5,8 @@ description = CLI Command Parser
 url = https://github.com/dskrypa/cli_command_parser
 project_urls =
     Source = https://github.com/dskrypa/cli_command_parser
+    Documentation = https://dskrypa.github.io/cli_command_parser
+    Issues = https://github.com/dskrypa/cli_command_parser/issues
 license = Apache 2.0
 license_files = LICENSE
 long_description = file: readme.rst
@@ -30,6 +32,7 @@ classifiers =
 
 [options]
 include_package_data = True
+entry_points = file: entry_points.txt
 packages = find:
 package_dir = = lib
 python_requires = >=3.8
